demo/requirements.txt
# Requirements for the Prompt Decorators OpenAI Demo
prompt-decorators
openai
python-dotenv
rich  # For pretty terminal output
typer  # For CLI interface
